refactor: Sidebar as functional component
