+++ frontend/src/Sidebar/Sidebar.js
@@ -1,67 +1,26 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { getTree, getChildren } from './utils';
+import buildTree from './utils';
 
 import { Wrapper } from './styled';
 
-/* eslint-disable camelcase */
-export default class Sidebar extends React.Component {
-  static propTypes = {
-    articles: PropTypes.array.isRequired,
-    match: PropTypes.object.isRequired,
-    onArticleSelect: PropTypes.func,
-  };
+let articlesHtml = [];
 
-  static defaultProps = {
-    onArticleSelect: () => {},
-  };
+const Sidebar = ({ articles, match, onArticleSelect }) => {
+  if (articlesHtml.length === 0) {
+    articlesHtml = buildTree(articles, match.params.id, onArticleSelect);
+  }
+  return <Wrapper>{articlesHtml}</Wrapper>;
+};
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      articlesHtml: [],
-    };
-  }
+Sidebar.propTypes = {
+  articles: PropTypes.array.isRequired,
+  match: PropTypes.object.isRequired,
+  onArticleSelect: PropTypes.func,
+};
 
-  componentDidMount() {
-    this.buildTree(this.props.articles, this.props.match.params.id);
-  }
+Sidebar.defaultProps = {
+  onArticleSelect: () => {},
+};
 
-  componentWillReceiveProps(nextProps) {
-    const { articles } = nextProps;
-    this.buildTree(articles, nextProps.match.params.id);
-  }
-  buildTree(articles, id) {
-    let selectedArticlePath = [];
-    const { onArticleSelect } = this.props;
-    // only build the sidebar tree once on startup, otherwise the tree is
-    // collapsed everytime an article is selected/viewed
-    if (this.state.articlesHtml.length === 0) {
-      if (id) {
-        const selectedArticles = articles.filter(article => article._id === id);
-        if (selectedArticles.length) {
-          const { topic, sub_topic, title } = selectedArticles[0];
-          selectedArticlePath = [topic.name, title];
-          // eslint-disable-next-line camelcase
-          if (sub_topic) {
-            selectedArticlePath = selectedArticlePath.concat(
-              sub_topic.name.split('>')
-            );
-          }
-        }
-      }
-      const tree = getTree(articles);
-      const articlesHtml = getChildren(
-        tree,
-        selectedArticlePath,
-        onArticleSelect
-      );
-      this.setState({ articlesHtml });
-    }
-  }
-
-  render() {
-    return <Wrapper>{this.state.articlesHtml}</Wrapper>;
-  }
-}
-/* eslint-enable camelcase */
+export default Sidebar;
